% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyHaplotyper.R
\docType{data}
\name{demo_ped}
\alias{demo_ped}
\title{pedigree}
\format{
a data.frame with 661 rows (samples) and
4 columns: genotype, mother, father, sample_nr
}
\description{
A data.frame with a pedigree; used in vignette, stored in
PolyHaplotyper_demo.RData
}
\keyword{datasets}
