% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyHaplotyper.R
\docType{data}
\name{demo_snpdos}
\alias{demo_snpdos}
\title{dosages of SNP alleles}
\format{
a data.frame with 30 rows (SNPs) and 663 columns. The first
two columns contain the SNP name and contig; the remaining columns
contain the allele dosages (integers in 0..6 or NA) of the SNPs in
661 samples
}
\description{
A data.frame with the allele dosages for 30 SNPs in 661
samples; used in vignette, stored in PolyHaplotyper_demo.RData
}
\keyword{datasets}
