% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyHaplotyper.R
\docType{data}
\name{phblocks}
\alias{phblocks}
\title{List of markers per haploblock}
\format{
a list with 2 character vectors
}
\description{
A list with for each haploblock the names of the markers it
contains; used in manuals, stored in PolyHaplotyper_small.RData
}
\keyword{datasets}
