% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyHaplotyper.R
\docType{data}
\name{phFS}
\alias{phFS}
\title{members of FS families}
\format{
a list with 4 vectors, each with the (here numeric) names of
the FS members
}
\description{
A list with for each FS family the names of the individuals it
contains; used in manuals, stored in PolyHaplotyper_small.RData
}
\keyword{datasets}
