% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyHaplotyper.R
\docType{data}
\name{phpar}
\alias{phpar}
\title{parents of FS families}
\format{
a matrix with 4 rows, 1 per FS family, and 2 columns, 1 for each
parent
}
\description{
A matrix with the names of the parents of each FS family;
used in manuals, stored in PolyHaplotyper_small.RData
}
\keyword{datasets}
