% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyHaplotyper.R
\docType{data}
\name{phped}
\alias{phped}
\title{pedigree}
\format{
a data.frame with 661 rows (samples) and
4 columns: genotype, mother, father, sample_nr
}
\description{
A data.frame with a pedigree; used in manuals, stored in
PolyHaplotyper_small.RData
}
\keyword{datasets}
