% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyHaplotyper.R
\docType{data}
\name{phresults}
\alias{phresults}
\title{haplotyping results}
\format{
a list with 2 elements, one per haploblock; each element itself
a list with multiple components
}
\description{
A list with the results of function inferHaplotypes; used in
manuals, stored in PolyHaplotyper_small.RData
}
\keyword{datasets}
