% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyHaplotyper.R
\name{run.SATlotyper}
\alias{run.SATlotyper}
\title{A simple interface to run SATlotyper}
\usage{
run.SATlotyper(path_to_SATlotyper, infile, outfile,
SAT_solver="sat4j.conf")
}
\arguments{
\item{path_to_SATlotyper}{path to the folder where SATlotyper.jar and
the *.conf files of the SAT-solvers are located}

\item{infile}{name (and path) to the SATlotyper input file}

\item{outfile}{name (and path) for the SATlotyper output file (an xml file)}

\item{SAT_solver}{name of the *.conf file for the SAT-solver to use.
Default "sat4j.conf" because this works under both Windows and Linux.}
}
\value{
The return is a list with elements:\cr
$cmd : the command passed to the system\cr
$result: screen output from SATlotyper; this
contains some extra info not present in the output xml file)\cr
The main result is the outfile
}
\description{
A simple interface to run SATlotyper
}
\details{
This function issues a system command to invoke SATlotyper.
Java and SATlotyper must be installed. This is just a simple interface for
convenience; for more control run SATlotyper directly from the command window.
}
